LPC84x / System - reverted the last conversion from CPP to C files
