+++ source/targets/LPC84x/lpc84x_interrupts.cpp
@@ -1,7 +1,7 @@
 // ----------------------------------------------------------------------------
 // @file    lpc84x_interrupts.cpp
 // @brief   IRQ handlers and vector table for NXP LPC84x MCU.
-// @date    13 June 2018
+// @date    21 June 2018
 // ----------------------------------------------------------------------------
 //
 // Xarmlib 0.1.0 - https://github.com/hparracho/Xarmlib
@@ -32,13 +32,6 @@
 #include "system/target"
 
 #ifdef __LPC84X__
-
-namespace xarmlib
-{
-namespace targets
-{
-namespace lpc84x
-{
 
 extern "C"
 {
@@ -293,9 +286,5 @@ void SysTick_Handler(void)
 
 
 } // extern "C"
-
-} // namespace lpc84x
-} // namespace targets
-} // namespace xarmlib
 
 #endif // __LPC84X__
